requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-openai
chromadb
pymupdfpymupdftiktoken